move `logging_cfg` to top-level package and reduce test timout
